pyproject.toml
[project]
name = "archydra"
version = "0.1.0"
description = ""
authors = [
    {name = "jcapelle13",email = "14522128+jcapelle13@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "click (>=8.3.1,<9.0.0)",
    "ruamel-yaml (>=0.19.1,<0.20.0)",
    "google-api-python-client (>=2.187.0,<3.0.0)",
    "google-auth-oauthlib (>=1.2.3,<2.0.0)",
    "google-auth-httplib2 (>=0.3.0,<0.4.0)"
]

[project.scripts]
archydra = "archydra:cli"

[tool.poetry]
packages = [{include = "archydra", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.14.10,<0.15.0)",
    "vermin (>=1.8.0,<2.0.0)",
]
